+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatCaret.java
@@ -18,6 +18,8 @@ package com.formdev.flatlaf.ui;
 
 import static com.formdev.flatlaf.FlatClientProperties.*;
 import java.awt.EventQueue;
+import java.awt.Insets;
+import java.awt.Rectangle;
 import java.awt.event.FocusEvent;
 import java.awt.event.MouseEvent;
 import javax.swing.JFormattedTextField;
@@ -59,6 +61,19 @@ public class FlatCaret
 			if( length > 0 )
 				setDot( length );
 		}
+	}
+
+	@Override
+	protected void adjustVisibility( Rectangle nloc ) {
+		JTextComponent c = getComponent();
+		if( c != null && c.getUI() instanceof FlatTextFieldUI ) {
+			Insets padding = ((FlatTextFieldUI)c.getUI()).getPadding();
+			if( padding != null ) {
+				nloc.x -= padding.left;
+				nloc.y -= padding.top;
+			}
+		}
+		super.adjustVisibility( nloc );
 	}
 
 	@Override