+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatPasswordFieldUI.java
@@ -16,34 +16,32 @@
 
 package com.formdev.flatlaf.ui;
 
-import java.awt.Color;
-import java.awt.Dimension;
 import java.awt.Graphics;
-import java.awt.Graphics2D;
+import java.awt.Shape;
 import java.awt.Toolkit;
-import java.awt.event.FocusListener;
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
-import java.beans.PropertyChangeEvent;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicBoolean;
+import javax.swing.Action;
+import javax.swing.ActionMap;
 import javax.swing.Icon;
 import javax.swing.JComponent;
 import javax.swing.LookAndFeel;
+import javax.swing.SwingUtilities;
 import javax.swing.UIManager;
 import javax.swing.plaf.ComponentUI;
-import javax.swing.plaf.basic.BasicPasswordFieldUI;
-import javax.swing.text.Caret;
+import javax.swing.text.DefaultEditorKit;
+import javax.swing.text.Element;
 import javax.swing.text.JTextComponent;
+import javax.swing.text.PasswordView;
+import javax.swing.text.View;
 import com.formdev.flatlaf.icons.FlatCapsLockIcon;
 import com.formdev.flatlaf.ui.FlatStyleSupport.Styleable;
-import com.formdev.flatlaf.util.HiDPIUtils;
 
 /**
  * Provides the Flat LaF UI delegate for {@link javax.swing.JPasswordField}.
  *
- * <!-- BasicPasswordFieldUI -->
+ * <!-- BasicTextFieldUI -->
  *
  * @uiDefault PasswordField.font					Font
  * @uiDefault PasswordField.background				Color
@@ -56,42 +54,32 @@ import com.formdev.flatlaf.util.HiDPIUtils;
  * @uiDefault PasswordField.inactiveForeground		Color	used if not enabled (yes, this is confusing; this should be named disabledForeground)
  * @uiDefault PasswordField.border					Border
  * @uiDefault PasswordField.margin					Insets
- * @uiDefault PasswordField.echoChar				character
  * @uiDefault PasswordField.caretBlinkRate			int		default is 500 milliseconds
  *
- * <!-- FlatPasswordFieldUI -->
+ * <!-- FlatTextFieldUI -->
  *
  * @uiDefault Component.minimumWidth				int
  * @uiDefault Component.isIntelliJTheme				boolean
  * @uiDefault PasswordField.placeholderForeground	Color
  * @uiDefault PasswordField.focusedBackground		Color	optional
- * @uiDefault PasswordField.showCapsLock			boolean
- * @uiDefault PasswordField.capsLockIcon			Icon
  * @uiDefault TextComponent.selectAllOnFocusPolicy	String	never, once (default) or always
  * @uiDefault TextComponent.selectAllOnMouseClick	boolean
+ *
+ * <!-- FlatPasswordFieldUI -->
+ *
+ * @uiDefault PasswordField.echoChar				character
+ * @uiDefault PasswordField.showCapsLock			boolean
+ * @uiDefault PasswordField.capsLockIcon			Icon
  *
  * @author Karl Tauber
  */
 public class FlatPasswordFieldUI
-	extends BasicPasswordFieldUI
+	extends FlatTextFieldUI
 {
-	@Styleable protected int minimumWidth;
-	protected boolean isIntelliJTheme;
-	private Color background;
-	@Styleable protected Color disabledBackground;
-	@Styleable protected Color inactiveBackground;
-	@Styleable protected Color placeholderForeground;
-	@Styleable protected Color focusedBackground;
 	@Styleable protected boolean showCapsLock;
 	protected Icon capsLockIcon;
 
-	private Color oldDisabledBackground;
-	private Color oldInactiveBackground;
-
-	private FocusListener focusListener;
 	private KeyListener capsLockListener;
-	private Map<String, Object> oldStyleValues;
-	private AtomicBoolean borderShared;
 	private boolean capsLockIconShared = true;
 
 	public static ComponentUI createUI( JComponent c ) {
@@ -99,10 +87,8 @@ public class FlatPasswordFieldUI
 	}
 
 	@Override
-	public void installUI( JComponent c ) {
-		super.installUI( c );
-
-		applyStyle( FlatStyleSupport.getStyle( c ) );
+	protected String getPropertyPrefix() {
+		return "PasswordField";
 	}
 
 	@Override
@@ -110,47 +96,24 @@ public class FlatPasswordFieldUI
 		super.installDefaults();
 
 		String prefix = getPropertyPrefix();
-		minimumWidth = UIManager.getInt( "Component.minimumWidth" );
-		isIntelliJTheme = UIManager.getBoolean( "Component.isIntelliJTheme" );
-		background = UIManager.getColor( prefix + ".background" );
-		disabledBackground = UIManager.getColor( prefix + ".disabledBackground" );
-		inactiveBackground = UIManager.getColor( prefix + ".inactiveBackground" );
-		placeholderForeground = UIManager.getColor( prefix + ".placeholderForeground" );
-		focusedBackground = UIManager.getColor( prefix + ".focusedBackground" );
+		Character echoChar = (Character) UIManager.get( prefix + ".echoChar" );
+		if( echoChar != null )
+			LookAndFeel.installProperty( getComponent(), "echoChar", echoChar );
+
 		showCapsLock = UIManager.getBoolean( "PasswordField.showCapsLock" );
 		capsLockIcon = UIManager.getIcon( "PasswordField.capsLockIcon" );
-
-		LookAndFeel.installProperty( getComponent(), "opaque", false );
-
-		MigLayoutVisualPadding.install( getComponent() );
 	}
 
 	@Override
 	protected void uninstallDefaults() {
 		super.uninstallDefaults();
 
-		background = null;
-		disabledBackground = null;
-		inactiveBackground = null;
-		placeholderForeground = null;
-		focusedBackground = null;
 		capsLockIcon = null;
-
-		oldDisabledBackground = null;
-		oldInactiveBackground = null;
-
-		oldStyleValues = null;
-		borderShared = null;
-
-		MigLayoutVisualPadding.uninstall( getComponent() );
 	}
 
 	@Override
 	protected void installListeners() {
 		super.installListeners();
-
-		// necessary to update focus border and background
-		focusListener = new FlatUIUtils.RepaintFocusListener( getComponent(), null );
 
 		// update caps lock indicator
 		capsLockListener = new KeyAdapter() {
@@ -168,7 +131,6 @@ public class FlatPasswordFieldUI
 			}
 		};
 
-		getComponent().addFocusListener( focusListener );
 		getComponent().addKeyListener( capsLockListener );
 	}
 
@@ -176,43 +138,24 @@ public class FlatPasswordFieldUI
 	protected void uninstallListeners() {
 		super.uninstallListeners();
 
-		getComponent().removeFocusListener( focusListener );
 		getComponent().removeKeyListener( capsLockListener );
-		focusListener = null;
 		capsLockListener = null;
 	}
 
 	@Override
-	protected Caret createCaret() {
-		return new FlatCaret( UIManager.getString( "TextComponent.selectAllOnFocusPolicy" ),
-			UIManager.getBoolean( "TextComponent.selectAllOnMouseClick" ) );
+	protected void installKeyboardActions() {
+		super.installKeyboardActions();
+
+		// map "select-word" action (double-click) to "select-line" action
+		ActionMap map = SwingUtilities.getUIActionMap( getComponent() );
+		if( map != null && map.get( DefaultEditorKit.selectWordAction ) != null ) {
+			Action selectLineAction = map.get( DefaultEditorKit.selectLineAction );
+			if( selectLineAction != null )
+				map.put( DefaultEditorKit.selectWordAction, selectLineAction );
+		}
 	}
 
 	@Override
-	protected void propertyChange( PropertyChangeEvent e ) {
-		String propertyName = e.getPropertyName();
-		if( "editable".equals( propertyName ) || "enabled".equals( propertyName ) )
-			updateBackground();
-		else
-			super.propertyChange( e );
-		FlatTextFieldUI.propertyChange( getComponent(), e, this::applyStyle );
-	}
-
-	/**
-	 * @since TODO
-	 */
-	protected void applyStyle( Object style ) {
-		oldDisabledBackground = disabledBackground;
-		oldInactiveBackground = inactiveBackground;
-
-		oldStyleValues = FlatStyleSupport.parseAndApply( oldStyleValues, style, this::applyStyleProperty );
-
-		updateBackground();
-	}
-
-	/**
-	 * @since TODO
-	 */
 	protected Object applyStyleProperty( String key, Object value ) {
 		if( key.equals( "capsLockIconColor" ) && capsLockIcon instanceof FlatCapsLockIcon ) {
 			if( capsLockIconShared ) {
@@ -222,24 +165,23 @@ public class FlatPasswordFieldUI
 			return ((FlatCapsLockIcon)capsLockIcon).applyStyleProperty( key, value );
 		}
 
-		if( borderShared == null )
-			borderShared = new AtomicBoolean( true );
-		return FlatStyleSupport.applyToAnnotatedObjectOrBorder( this, key, value, getComponent(), borderShared );
+		return super.applyStyleProperty( key, value );
 	}
 
-	private void updateBackground() {
-		FlatTextFieldUI.updateBackground( getComponent(), background,
-			disabledBackground, inactiveBackground,
-			oldDisabledBackground, oldInactiveBackground );
+	@Override
+	public View create( Element elem ) {
+		return new PasswordView( elem );
 	}
 
 	@Override
 	protected void paintSafely( Graphics g ) {
-		FlatTextFieldUI.paintBackground( g, getComponent(), isIntelliJTheme, focusedBackground );
-		FlatTextFieldUI.paintPlaceholder( g, getComponent(), placeholderForeground );
-		paintCapsLock( g );
+		// safe and restore clipping area because super.paintSafely() modifies it
+		// and the caps lock icon would be truncated
+		Shape oldClip = g.getClip();
+		super.paintSafely( g );
+		g.setClip( oldClip );
 
-		super.paintSafely( HiDPIUtils.createGraphicsTextYCorrection( (Graphics2D) g ) );
+		paintCapsLock( g );
 	}
 
 	protected void paintCapsLock( Graphics g ) {
@@ -254,20 +196,5 @@ public class FlatPasswordFieldUI
 		int y = (c.getHeight() - capsLockIcon.getIconHeight()) / 2;
 		int x = c.getWidth() - capsLockIcon.getIconWidth() - y;
 		capsLockIcon.paintIcon( c, g, x, y );
-	}
-
-	@Override
-	protected void paintBackground( Graphics g ) {
-		// background is painted elsewhere
-	}
-
-	@Override
-	public Dimension getPreferredSize( JComponent c ) {
-		return FlatTextFieldUI.applyMinimumWidth( c, super.getPreferredSize( c ), minimumWidth );
-	}
-
-	@Override
-	public Dimension getMinimumSize( JComponent c ) {
-		return FlatTextFieldUI.applyMinimumWidth( c, super.getMinimumSize( c ), minimumWidth );
 	}
 }